requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
sqlalchemy==2.0.36
openai>=1.30.0
python-dotenv==1.0.1
pydantic-settings==2.7.0
python-multipart==0.0.20
aiofiles==24.1.0
httpx==0.28.1
